rename `strains.py` to `strain.py`
