+++ src/nplinker/genomics/antismash/antismash_loader.py
@@ -5,7 +5,7 @@ from Bio import SeqIO
 from Bio import SeqRecord
 from nplinker.genomics import BGC
 from nplinker.logconfig import LogConfig
-from nplinker.strains import Strain
+from nplinker.strain import Strain
 from nplinker.utils import list_dirs
 from nplinker.utils import list_files
 from ..abc import BGCLoaderBase
