+++ src/nplinker/scoring/linking/data_links.py
@@ -14,7 +14,7 @@ from .utils import isinstance_all
 
 if TYPE_CHECKING:
     from nplinker.strain_collection import StrainCollection
-    from nplinker.strains import Strain
+    from nplinker.strain import Strain
 
 logger = LogConfig.getLogger(__name__)
