+++ src/nplinker/strain.py
