+++ tests/conftest.py
@@ -3,7 +3,7 @@ from nplinker.globals import STRAIN_MAPPINGS_FILENAME
 from nplinker.metabolomics import Spectrum
 from nplinker.metabolomics.metabolomics import load_spectra
 from nplinker.strain_collection import StrainCollection
-from nplinker.strains import Strain
+from nplinker.strain import Strain
 from . import DATA_DIR
 from . import GNPS_DATA_DIR
