+++ tests/genomics/test_bgc.py
@@ -1,6 +1,6 @@
 from nplinker.genomics import BGC
 from nplinker.genomics import GCF
-from nplinker.strains import Strain
+from nplinker.strain import Strain
 
 
 def test_default():
