+++ tests/test_strain_loader.py
@@ -1,7 +1,7 @@
 import json
 import pytest
 from nplinker.strain_loader import load_user_strains
-from nplinker.strains import Strain
+from nplinker.strain import Strain
 
 
 @pytest.fixture
